moved download activity to provider, added exception for when field missing from cursor

+++ src/com/axelby/podax/DBAdapter.java
@@ -433,7 +433,7 @@ public class DBAdapter {
 		return collectPodcasts(c).get(0);
 	}
 	
-	public void updateActiveDownloadId(Integer podcastId) {
+	public void updateActiveDownloadId(Long podcastId) {
 		ContentValues values = new ContentValues();
 		values.put("activeDownloadId", podcastId);
 		_db.update("podax", values, "", null);
